# swagger.yaml
openapi: 3.0.0
info:
  title: MindLoo API
  version: 2.0.0
  description: |
    Complete RESTful API for managing patients, therapists, appointments, and therapy-related operations.
    
    ## Features
    - Patient and Therapist Management
    - Appointment Scheduling
    - Therapy Session Tracking
    - Medical Records Management
    - Secure Messaging System
    - Review and Rating System
    - Treatment Planning and Goal Tracking
    
  contact:
    name: Stefan Ramač
    url: https://www.linkedin.com/in/stefanramac
servers:
  - url: http://localhost:3000
    description: Development server
...
components:
  schemas:
...
    MedicalRecord:
      type: object
      required:
        - patientId
        - recordType
        - title
        - addedBy
      properties:
        recordId:
          type: string
          readOnly: true
        patientId:
          type: string
        recordType:
          type: string
          enum: [diagnosis, medication, allergy, lab-result, history, other]
        title:
          type: string
        description:
          type: string
        diagnosis:
          type: object
          properties:
            code:
              type: string
              example: "F41.1"
            name:
              type: string
              example: "Generalized Anxiety Disorder"
            severity:
              type: string
              enum: [mild, moderate, severe]
        medications:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              dosage:
                type: string
              frequency:
                type: string
              startDate:
                type: string
                format: date
              endDate:
                type: string
                format: date
        allergies:
          type: array
          items:
            type: object
            properties:
              allergen:
                type: string
              reaction:
                type: string
              severity:
                type: string
                enum: [mild, moderate, severe, life-threatening]
        addedBy:
          type: string
        isActive:
          type: boolean
          default: true
    
    # Message Schema
    Message:
      type: object
      required:
        - senderId
        - senderType
        - receiverId
        - receiverType
        - content
      properties:
        messageId:
          type: string
          readOnly: true
        conversationId:
          type: string
        senderId:
          type: string
        senderType:
          type: string
          enum: [patient, therapist]
        receiverId:
          type: string
        receiverType:
          type: string
          enum: [patient, therapist]
        subject:
          type: string
        content:
          type: string
        messageType:
          type: string
          enum: [text, appointment-request, prescription, document, emergency]
          default: text
        priority:
          type: string
          enum: [low, normal, high, urgent]
          default: normal
        isRead:
          type: boolean
          default: false
    
    # Review Schema
...
paths:
...
  /api/medical-records:
    post:
      tags:
        - Medical Records
      summary: Create a new medical record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecord'
      responses:
        '201':
          description: Medical record created successfully
  /api/medical-records/patient/{patientId}:
    get:
      tags:
        - Medical Records
      summary: Get all medical records for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
        - name: recordType
          in: query
          schema:
            type: string
            enum: [diagnosis, medication, allergy, lab-result, history, other]
        - name: isActive
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of medical records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MedicalRecord'
  /api/medical-records/{id}:
    get:
      tags:
        - Medical Records
      summary: Get medical record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medical record data
        '404':
          description: Record not found
    
    patch:
      tags:
        - Medical Records
      summary: Update medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MedicalRecord'
      responses:
        '200':
          description: Medical record updated successfully
    
    delete:
      tags:
        - Medical Records
      summary: Delete a medical record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Medical record deleted successfully
  /api/medical-records/patient/{patientId}/summary:
    get:
      tags:
        - Medical Records
      summary: Get medical summary for a patient
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient medical summary
  
  # Messages Endpoints
  /api/messages:
    post:
      tags:
        - Messages
      summary: Send a new message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Message'
      responses:
        '201':
          description: Message sent successfully
  /api/messages/conversation/{conversationId}:
    get:
      tags:
        - Messages
      summary: Get all messages in a conversation
      parameters:
        - name: conversationId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
  /api/messages/user/{userId}:
    get:
      tags:
        - Messages
      summary: Get all conversations for a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: List of conversations
  /api/messages/{id}:
    get:
      tags:
        - Messages
      summary: Get message by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message data
        '404':
          description: Message not found
    
    delete:
      tags:
        - Messages
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message deleted successfully
  /api/messages/{id}/read:
    patch:
      tags:
        - Messages
      summary: Mark message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Message marked as read
  
  # Reviews Endpoints
...